requirements.txt
Pillow==6.0.0
image==1.5.27
opencv-python==4.0.0.21